Models folder created and module models added
