+++ lib/models/colours_model.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class Colours {
+  final String name;
+  final String jpgAsset;
+  final Color bgColor;
+  final Color fontColor;
+
+  Colours({
+    required this.name,
+    required this.jpgAsset,
+    required this.bgColor,
+    required this.fontColor,
+  });
+}
