+++ lib/models/itemdata_model.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class ItemData {
+  final String iconAsset;
+  final String title;
+  final String description;
+  final Color backgroundColor;
+
+  ItemData({
+    required this.iconAsset,
+    required this.title,
+    required this.description,
+    required this.backgroundColor,
+  });
+}
