+++ lib/models/occupation_model.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class Occupation {
+  final String name;
+  final String description;
+  final String svgAsset;
+  final Color backgroundColor;
+
+  Occupation({
+    required this.name,
+    required this.description,
+    required this.svgAsset,
+    required this.backgroundColor,
+  });
+}
