+++ lib/models/season_model.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class Season {
+  final String name;
+  final String description;
+  final String imageAsset;
+  final Color backgroundColor;
+
+  Season({
+    required this.name,
+    required this.description,
+    required this.imageAsset,
+    required this.backgroundColor,
+  });
+}
