+++ lib/pages/modules/atoz.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'dart:async';
 import 'package:flutter_tts/flutter_tts.dart';
+import 'package:learn/models/itemdata_model.dart';
 
 class ItemTile extends StatelessWidget {
   final int index;
@@ -220,20 +221,6 @@ class _PopupDialogState extends State<_PopupDialog> {
       ),
     );
   }
-}
-
-class ItemData {
-  final String iconAsset;
-  final String title;
-  final String description;
-  final Color backgroundColor;
-
-  ItemData({
-    required this.iconAsset,
-    required this.title,
-    required this.description,
-    required this.backgroundColor,
-  });
 }
 
 class AtoZ extends StatefulWidget {
