+++ lib/pages/modules/seasons.dart
@@ -1,23 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'package:learn/models/season_model.dart';
 import 'package:learn/utils/constants.dart';
 
-class Season {
-  final String name;
-  final String description;
-  final String imageAsset;
-  final Color backgroundColor;
-
-  Season({
-    required this.name,
-    required this.description,
-    required this.imageAsset,
-    required this.backgroundColor,
-  });
-}
-
 class SeasonsPage extends StatelessWidget {
-  SeasonsPage({Key? key}) : super(key: key);
+  const SeasonsPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -86,7 +73,7 @@ class SeasonPopup extends StatefulWidget {
   final int currentIndex;
   final List<Season> seasons;
 
-  SeasonPopup({
+  const SeasonPopup({
     required this.currentIndex,
     required this.seasons,
   });
